Solution / Naming corrections
