+++ DCaPST.sln
@@ -1,22 +1,41 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 14
-VisualStudioVersion = 14.0.25420.1
+# Visual Studio 15
+VisualStudioVersion = 15.0.28010.2048
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DCaPST", "DCaPST.csproj", "{96037C21-E821-4CF8-98B5-C8BD2446F4E0}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "DcapsTest", "..\DcapsTest\DcapsTest\DcapsTest.csproj", "{7930C660-2206-4215-B9CA-EE7C65A006E0}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
+		Debug|x86 = Debug|x86
 		Release|Any CPU = Release|Any CPU
+		Release|x86 = Release|x86
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Debug|x86.Build.0 = Debug|Any CPU
 		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Release|x86.ActiveCfg = Release|x86
+		{96037C21-E821-4CF8-98B5-C8BD2446F4E0}.Release|x86.Build.0 = Release|x86
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Debug|x86.Build.0 = Debug|Any CPU
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Release|Any CPU.Build.0 = Release|Any CPU
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Release|x86.ActiveCfg = Release|Any CPU
+		{7930C660-2206-4215-B9CA-EE7C65A006E0}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {F36D6CE1-25BA-45EF-8BCD-72FEA02DC4D4}
 	EndGlobalSection
 EndGlobal
